Refactor display name assignment in DisplayNameHistory and clean up unused code in ProfileOptions and LatencyHistory
